This is a fashion blog.

+++ src/main/java/com/boluwatifeproj/fashionblog/service/PostServiceImpl.java
@@ -13,57 +13,72 @@ import org.springframework.web.bind.annotation.RequestBody;
 import java.util.List;
 import java.util.Optional;
 
+import java.util.List;
+import java.util.Optional;
+
 @Service
-public class PostServiceImpl {
+public class PostServiceImpl implements PostService {
     private PostRepository postRepository;
     @Autowired
     public PostServiceImpl(PostRepository postRepository) {
         this.postRepository = postRepository;
     }
-    public ResponseEntity<Post> savePost(Post post){
-       postRepository.save(post);
-       return new ResponseEntity<>(post, HttpStatus.CREATED);
+
+    @Override
+    public ResponseEntity<Post> savePost(Post post) {
+        return new ResponseEntity<>(post, HttpStatus.CREATED);
     }
 
+    @Override
     public ResponseEntity<List<Post>> getAllPosts() {
         List<Post> postList = postRepository.findAll();
-       return new ResponseEntity<>(postList, HttpStatus.FOUND);
-
+        return new ResponseEntity<>(postList, HttpStatus.FOUND);
     }
-    public ResponseEntity<Post> editPostById (Long id, @RequestBody Post newPost){
+
+    @Override
+    public ResponseEntity<Post> editPostById(Long id, Post newPost) {
         Optional<Post> post = postRepository.findById(id);
         if (post.isPresent()){
             Post post1 = post.get();
             post1.setTitle(newPost.getTitle());
             post1.setContent(newPost.getContent());
-            postRepository.save(post1);
-            return new ResponseEntity<Post>(post1, HttpStatus.OK);
+            return new ResponseEntity<>(post1, HttpStatus.CREATED);
         }else{
-          return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
+    @Override
     public ResponseEntity<Post> getPostById(Long id) {
         Optional<Post> post = postRepository.findById(id);
         if (post.isPresent()){
             Post post1 = post.get();
             return new ResponseEntity<>(post1, HttpStatus.FOUND);
+
         }else{
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
-    public ResponseEntity<String> deletePostById(Long id) throws Throwable {
-        postRepository.findById(id).orElseThrow(()-> new PostNotFoundException("POST NOT FOUND"));
-        postRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-    public ResponseEntity<List<Post>> getPostByTitle(String title) {
-        List<Post> postList = postRepository.findPostIgnoringCase(title);
-        if (postList.isEmpty()){
-            return new ResponseEntity<>(postList, HttpStatus.NO_CONTENT);
 
+    @Override
+    public ResponseEntity<String> deletePostById(Long id) throws Throwable {
+        Optional<Post> post = postRepository.findById(id);
+        if (post.isPresent()){
+            Post post1 = post.get();
+            postRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
         }else{
-            return new ResponseEntity<>(postList, HttpStatus.FOUND);
+            throw new PostNotFoundException("Post not found");
+        }
+    }
+
+    @Override
+    public ResponseEntity<List<Post>> getPostByTitle(String title) {
+        List <Post> postList = postRepository.findByTitle(title);
+        if (postList.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }else{
+            return new ResponseEntity<>(HttpStatus.FOUND);
         }
     }
 }
